+++ examples/coagents-routing/agent-js/src/agent.ts
@@ -6,9 +6,9 @@ import { tool } from "@langchain/core/tools";
 import { z } from "zod";
 import { RunnableConfig } from "@langchain/core/runnables";
 import {
-  copilotKitCustomizeConfig,
-  copilotKitExit,
-} from "@copilotkit/sdk-js/langchain";
+  copilotkitCustomizeConfig,
+  copilotkitExit,
+} from "@copilotkit/sdk-js/langgraph";
 import { AIMessage, HumanMessage, ToolMessage } from "@langchain/core/messages";
 import { getModel } from "./model";
 import { END, MemorySaver, StateGraph } from "@langchain/langgraph";
@@ -27,7 +27,7 @@ export async function email_node(state: AgentState, config: RunnableConfig) {
    * Write an email.
    */
 
-  const modifiedConfig = copilotKitCustomizeConfig(config, {
+  const modifiedConfig = copilotkitCustomizeConfig(config, {
     emitToolCalls: true,
   });
 
@@ -60,7 +60,7 @@ export async function send_email_node(
    * Send an email.
    */
 
-  await copilotKitExit(config);
+  await copilotkitExit(config);
 
   const lastMessage = state.messages[state.messages.length - 1] as ToolMessage;
   const content =