+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,21 @@
 # ui
 
+## 1.5.12-next.1
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.5.12-next.1
+- @copilotkit/react-core@1.5.12-next.1
+- @copilotkit/shared@1.5.12-next.1
+
+## 1.5.12-next.0
+
+### Patch Changes
+
+- @copilotkit/react-core@1.5.12-next.0
+- @copilotkit/runtime-client-gql@1.5.12-next.0
+- @copilotkit/shared@1.5.12-next.0
+
 ## 1.5.11
 
 ### Patch Changes